requirements.txt
# Core dependencies
setuptools>=68.0,<81.0  # python-Wappalyzer still imports pkg_resources
fastapi==0.115.6
httpx==0.28.1
pydantic==2.10.5
pydantic-settings==2.7.1
apscheduler==3.10.4
tldextract==5.1.3
beautifulsoup4==4.12.3
tenacity==9.0.0
python-Wappalyzer==0.3.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-mock==3.14.0

# Linting / formatting
ruff==0.9.10
